requirements.txt
pandas==2.0.3
psycopg2==2.9.9
python-decouple==3.8
Requests==2.32.3
